requirements.txt
# Core bioinformatics libraries
biopython>=1.80
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
pandas>=1.5.0

# Sequence alignment
pairwise2>=1.0.0

# File I/O and data processing
h5py>=3.7.0
scipy>=1.10.0

# CLI and utilities
click>=8.0.0
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=22.0.0
flake8>=5.0.0